requirements.txt
paddlepaddle==2.4.2
matplotlib
